Implement downloading via MediaStore on Android > 32, reduce permissions

+++ app/src/main/kotlin/files/DownloadFileListener.kt
@@ -25,37 +25,30 @@ open class DownloadFileListener {
                             downloadOrInstallId: Long, type: DownloadType) {
         val path = Downloader.getNormalDownloadPath(context, downloadOrInstallId)
         val file = File(path, fileName)
-        Mitch.externalFileManager.moveToDownloads(context, file) { newFileName ->
-            path.deleteRecursively()
-            if (newFileName == null) {
-                Toast.makeText(context, R.string.dialog_missing_file_title, Toast.LENGTH_LONG)
-                    .show()
-                return@moveToDownloads
-            }
-            Mitch.externalFileManager.doGetViewIntent(context, newFileName) { viewIntent ->
-                if (viewIntent == null) {
-                    Toast.makeText(context, R.string.dialog_missing_file_title, Toast.LENGTH_LONG)
-                        .show()
-                    return@doGetViewIntent
-                }
-                val intent = Intent.createChooser(viewIntent, context.resources.getString(R.string.select_app_for_file))
-                val pendingIntent = PendingIntentCompat.getActivity(
-                    context, 0, intent,
-                    PendingIntent.FLAG_UPDATE_CURRENT, false
-                )!!
+        val newUri = Mitch.externalFileManager.doMoveToDownloads(context, file)
+        if (newUri == null) {
+            Toast.makeText(context, R.string.dialog_missing_file_title, Toast.LENGTH_LONG)
+                .show()
+            return
+        }
+        path.deleteRecursively()
+        val viewIntent = Mitch.externalFileManager.getViewIntent(newUri)
+        val intent = Intent.createChooser(viewIntent, context.resources.getString(R.string.select_app_for_file))
+        val pendingIntent = PendingIntentCompat.getActivity(
+            context, 0, intent,
+            PendingIntent.FLAG_UPDATE_CURRENT, false
+        )!!
 
-                NotificationCompat.Builder(context, NOTIFICATION_CHANNEL_ID_INSTALL_NEEDED).apply {
-                    setSmallIcon(R.drawable.ic_mitch_notification)
-                    setContentTitle(context.resources.getString(R.string.notification_download_complete_title))
-                    setContentText(newFileName)
+        NotificationCompat.Builder(context, NOTIFICATION_CHANNEL_ID_INSTALL_NEEDED).apply {
+            setSmallIcon(R.drawable.ic_mitch_notification)
+            setContentTitle(context.resources.getString(R.string.notification_download_complete_title))
+            setContentText(newUri.lastPathSegment)
 
-                    priority = NotificationCompat.PRIORITY_HIGH
-                    setContentIntent(pendingIntent)
-                    setAutoCancel(true)
+            priority = NotificationCompat.PRIORITY_HIGH
+            setContentIntent(pendingIntent)
+            setAutoCancel(true)
 
-                    this@DownloadFileListener.notify(context, downloadOrInstallId, this.build())
-                }
-            }
+            this@DownloadFileListener.notify(context, downloadOrInstallId, this.build())
         }
     }
